pyproject.toml
[project]
name = "predictive-maintenance-dash"
version = "0.1.0"
description = "Dashboard de Mantenimiento Predictivo para activos industriales"
authors = [{name = "Jaime Campillay"}]
dependencies = [
    "polars>=1.0.0",
    "pandas",
    "sqlalchemy",
    "psycopg2-binary",
    "dash",
    "dash-bootstrap-components",
    "gunicorn",
    "plotly",
    "python-dotenv",
    "pydantic-settings",
    "fastapi>=0.133.1",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
